Fix Betty Style in all files and adding free_loop archive Adding env parameter for 'WARNING: externs should be avoided in .c files' --shell.c --shell.h --built_in.c --execute_line.c Adding comments to functions in all files Correcting indentation problems in all files Correction trailing white space to all files Eliminating unused functions --useful_fuc.c --shell.h Adding the file free_loop.c with the purpose of not exceeding 5 functions in useful_func.c
