+++ _which.c
@@ -1,25 +1,29 @@
 #include "shell.h"
-char *_which(char *arg[])
+/**
+*_which - fuction that looks for files in the current PATH.
+*@commands:the split arguments
+*Return: full_path of the command or NULL if it can´t find it
+*/
+char *_which(char **commands)
 {
 	list_p *head;
 	char *full_path = NULL;
 	struct stat st;
 
 	head = list_path();
-	while(head)
-        {
-                full_path = _strcat(head->dir, "/");
-                full_path = _strcat(full_path, arg[0]);
+	while (head)
+	{
+		full_path = _strcat(head->dir, "/");
+		full_path = _strcat(full_path, commands[0]);
 
-                if (stat(full_path, &st) == 0)
-                        break;
-                head = head->next;
-        }
+		if (stat(full_path, &st) == 0)
+			break;
+		head = head->next;
+	}
 	if (full_path == NULL)
 	{
 		perror("");
 		return (NULL);
 	}
-	/*free(head);*/
 	return (full_path);
 }
